Main Container and React Router
