+++ client/index.html
@@ -7,7 +7,7 @@
       href="https://fonts.googleapis.com/css?family=Assistant&display=swap"
       rel="stylesheet"
     />
-    <title>Document</title>
+    <title>Mood Ring</title>
   </head>
   <body>
     <div id="root"></div>
