+++ package.json
@@ -35,10 +35,11 @@
     "@babel/preset-react": "^7.8.3",
     "babel": "^6.23.0",
     "babel-loader": "^8.0.6",
+    "css-loader": "^3.4.2",
     "html-webpack-plugin": "^3.2.0",
     "nodemon": "^2.0.2",
+    "react-router-dom": "^5.1.2",
     "style-loader": "^1.1.3",
-    "css-loader": "^3.4.2",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.10.1"
